+++ src/console.ts
@@ -32,6 +32,11 @@ const deleteSampleScores = (): void => {
   setAllScores(newScores)
 }
 
+const deleteInvalidScores = function () {
+  const validScores = getAllScores().filter((s: ScoreItem) => s.play.moves > 0)
+  setAllScores(validScores)
+}
+
 export default {
   getAllScores,
   setAllScores,
@@ -39,4 +44,5 @@ export default {
   deleteScoresByUser,
   addSampleScores,
   deleteSampleScores,
+  deleteInvalidScores,
 }